+++ crates/miden-testing/src/kernel_tests/tx/test_fee.rs
@@ -33,7 +33,7 @@ async fn create_account_with_fees() -> anyhow::Result<()> {
         .context("failed to execute account-creating transaction")?;
 
     let expected_fee = tx.compute_fee();
-    assert_eq!(expected_fee, tx.fee().amount());
+    assert_eq!(expected_fee, tx.fee().amount().inner());
 
     // We expect that the new account contains the note_amount minus the paid fee.
     let added_asset = FungibleAsset::new(chain.native_asset_id(), note_amount)?.sub(tx.fee())?;